Implemented a simple but extremely effective optimization.
When visiting rules in the ATN with a given token stream position we always get the same token end positions. So it makes sense to cache those and return cached values for all invocations of the same rule index and token index pair. This optimization brought down a 9.9M states walk to 24K.

- Added a new C++ code completion test case with a non-trivial code file.
- Updated the documentation for the change.
- Improved package.json to better deal witht he upcoming breaking change in npm@5 (prebulish script only called when publishing). Since the module only ships js files there is no need for an installation step.
- Tests now run directly on the transpiled code instead of letting mocha transpile first (and run on the TS code)
- New minor release + version number increase.