Optimised token retrieval, fixes bug #18
When filling the token list in the code completion engine we can use the simple `get` function, instead of the lookahead function `LT` which has a loop to get a token.
While changing that a bug was fixed where an additional candidate was returned at the beginning of a white space token.