Account for empty and fully-optional-body rules when collecting tokens
Currently, collectCandidates does not return all possible tokens when it encounters rules that do not necessarily consume tokens. This is due to two reasons:
1) collectFollowSets does not proceed to states beyond the end of a rule transitioned to. This means that subsequent tokens within the rule are not collected even if the rule transition is to a rule with fully-optional tokens or no tokens at all.
2) processRule always returns an empty-set RuleEndStatus when at the caret position. This means that tokens from subsequent transitions are not collected, even if the processed rule could reach the rule end state without consuming tokens.

This PR fixes both issues by adding property isExhaustive to the follow sets object to represent whether the collected follow sets represent all possible tokens that can follow, or if subsequent transitions need to be followed to further collect tokens, and using that to further collect follow sets within the processed rule after the rule transition and after the processed rule. Also, it removes the use of the epsilon token when collecting follow sets and refactors code to use isExhaustive property instead, ensuring that the epsilon token is never returned in the candidate tokens list.

Fixes mike-lischke/antlr4-c3#23