Fix: (#80) Candidate token's follow TokenList is inaccurate when the candidate's origin is ambiguous
When setting the follow list of a token suggestion, if the candidate token is already in the suggestions map we need to set the follow list to the longest common prefix of the current follow list and the newly computed one.

This does not change the treatment of preferred rules: if a token suggestion is masked by a preferred rule, it does not impact the follow set of the same token suggestion collected from a non-preferred rule.